refactor: schemas
