+++ api/types/common-application-card/index.ts
