+++ api/types/common-news-card/index.ts
