+++ api/types/page-element-basics/index.ts
@@ -1,0 +1,1 @@
+export * from './.type/index.js'
