+++ api/types/page-element-events/index.ts
@@ -1,0 +1,1 @@
+export * from './.type/index.js'
