+++ api/types/page-element-layout/index.ts
@@ -1,0 +1,1 @@
+export * from './.type/index.js'
