+++ api/types/page-element-news/index.ts
@@ -1,0 +1,1 @@
+export * from './.type/index.js'
