+++ api/types/portal-config-datasets/schema.js
@@ -21,7 +21,7 @@ export default {
   unevaluatedProperties: false,
   required: ['card', 'page'],
   properties: {
-    card: { $ref: 'https://github.com/data-fair/portals/portal-config-dataset-card' },
+    card: { $ref: 'https://github.com/data-fair/portals/common-dataset-card' },
     page: { $ref: 'https://github.com/data-fair/portals/portal-config-dataset-page' }
   }
 }
